requirements.txt
numpy==1.21.5
pandas==1.3.5
Pillow==9.0.1
streamlit==1.5.1
os
datetime
